+++ web/src/en.js
@@ -796,6 +796,7 @@ export default {
       unmuteAll: 'Unmute messages and mentions',
       muteIndicatorRegular: 'All messages muted, mentions still visible',
       muteIndicatorAll: 'All messages and mentions muted',
+      search: 'Search messages',
     },
     status: {
       cleared: 'Local message history cleared.',