+++ web/src/ru.js
@@ -796,6 +796,7 @@ export default {
       unmuteAll: 'Снять mute всех сообщений и упоминаний',
       muteIndicatorRegular: 'Mute всех сообщений, упоминания оставлены',
       muteIndicatorAll: 'Mute всех сообщений и упоминаний',
+      search: 'Поиск сообщений',
     },
     status: {
       cleared: 'Локальная история сообщений очищена.',